third_party/blink/web_tests/platform/generic/idle-callback/test-runner-run-idle-tasks-expected.txt

Tests that testRunner.runIdleTasks will force idle tasks to run.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS x is false
PASS x is true
PASS successfullyParsed is true

TEST COMPLETE

Hello, world!
